refactor: split AssignmentDefinitionController into folder-based facade
- Create 7 focused sub-classes under AssignmentDefinition/ folder
- Thin facade index.js preserves public API contract
- All sub-classes follow constructor injection pattern
- GAS-safe: no require() calls in production code
- Update test globals to pre-load sub-classes
- Section 8 complete in ACTION_PLAN_DATE_HARDENING.md
